refactor(component): Step3에서의 재사용을 위해 Step2RightSideComponent의 시험지 요약 부분 컴포넌트 분리 (#116)
- Step2RightSideComponent로부터 ExamSummaryComponent 분리
